+++ client/src/components/pages/header-footer/Footer.jsx
@@ -27,7 +27,7 @@ const Footer = () => {
             rel="noopener noreferrer"
             href="https://github.com/Ojself/cyberwarfare4k/blob/master/changelog.md"
           >
-            0.1.3
+            0.1.4
           </a>{" "}
         </div>
       </div>